requirements.txt
python-telegram-bot[job-queue] == 20.7
requests
python-dotenv
aiohttp
